tests/test_production.py
"""
Production test suite
"""
import pytest
import asyncio
from httpx import AsyncClient
from unittest.mock import Mock, patch
from src.main_production import app
from src.models.database import Base, engine

@pytest.fixture
async def client():
    """Create test client"""
    async with AsyncClient(app=app, base_url="http://test") as ac:
        yield ac

@pytest.fixture
def auth_headers():
    """Test authentication headers"""
    return {"Authorization": "Bearer test-api-key"}

class TestEmailProcessing:
    """Test email processing endpoints"""
    
    @pytest.mark.asyncio
    async def test_process_email_success(self, client, auth_headers):
        """Test successful email processing"""
        response = await client.post(
            "/api/v1/emails/process",
            json={
                "sender": "test@example.com",
                "subject": "Test Email",
                "content": "This is a test email"
            },
            headers=auth_headers
        )
        
        assert response.status_code == 200
        data = response.json()
        assert data["status"] == "processing"
        assert "id" in data
    
    @pytest.mark.asyncio
    async def test_rate_limiting(self, client, auth_headers):
        """Test rate limiting"""
        # Send many requests
        for i in range(105):  # Exceed limit
            response = await client.post(
                "/api/v1/emails/process",
                json={
                    "sender": f"test{i}@example.com",
                    "subject": "Test",
                    "content": "Test"
                },
                headers=auth_headers
            )
            
            if i >= 100:  # Should be rate limited
                assert response.status_code == 429
                assert "rate limit" in response.json()["error"].lower()
    
    @pytest.mark.asyncio
    async def test_batch_processing(self, client, auth_headers):
        """Test batch email processing"""
        emails = [
            {
                "sender": f"user{i}@example.com",
                "subject": f"Email {i}",
                "content": f"Content {i}"
            }
            for i in range(10)
        ]
        
        response = await client.post(
            "/api/v1/emails/batch",
            json={
                "emails": emails,
                "async_processing": True
            },
            headers=auth_headers
        )
        
        assert response.status_code == 200
        data = response.json()
        assert data["count"] == 10
        assert len(data["task_ids"]) == 10

class TestWebSocket:
    """Test WebSocket functionality"""
    
    @pytest.mark.asyncio
    async def test_websocket_connection(self, client):
        """Test WebSocket connection"""
        async with client.websocket_connect("/ws/test-org/test-user") as websocket:
            data = await websocket.receive_json()
            assert data["type"] == "connection"
            assert data["status"] == "connected"
    
    @pytest.mark.asyncio
    async def test_websocket_updates(self, client):
        """Test real-time updates"""
        async with client.websocket_connect("/ws/test-org/test-user") as websocket:
            # Wait for connection message
            await websocket.receive_json()
            
            # Send ping
            await websocket.send_json({"type": "ping"})
            
            # Receive pong
            pong = await websocket.receive_json()
            assert pong["type"] == "pong"

class TestSecurity:
    """Test security features"""
    
    @pytest.mark.asyncio
    async def test_invalid_api_key(self, client):
        """Test invalid API key rejection"""
        response = await client.post(
            "/api/v1/emails/process",
            json={"sender": "test@example.com", "subject": "Test", "content": "Test"},
            headers={"Authorization": "Bearer invalid-key"}
        )
        
        assert response.status_code == 401
    
    @pytest.mark.asyncio
    async def test_ip_whitelist(self, client):
        """Test IP whitelist enforcement"""
        # Test with non-whitelisted IP
        # Implementation depends on your setup
        pass

class TestPerformance:
    """Performance tests"""
    
    @pytest.mark.asyncio
    async def test_concurrent_requests(self, client, auth_headers):
        """Test handling concurrent requests"""
        async def make_request(i):
            return await client.post(
                "/api/v1/emails/process",
                json={
                    "sender": f"test{i}@example.com",
                    "subject": f"Test {i}",
                    "content": "Concurrent test"
                },
                headers=auth_headers
            )
        
        # Send 50 concurrent requests
        tasks = [make_request(i) for i in range(50)]
        responses = await asyncio.gather(*tasks)
        
        # All should succeed (within rate limits)
        success_count = sum(1 for r in responses if r.status_code == 200)
        assert success_count >= 45  # Allow some rate limiting

# Run tests
if __name__ == "__main__":
    pytest.main([__file__, "-v"])
